// src/app/models/featuring.ts
export class Featurer {
    lyric: String = null;
    singer: String = null;
    one: String = null;
    two: String = null;
    three: String = null;
    four: String = null;
    five: String = null;
    six: String = null;
    seven: String = null;
    eight: String = null;
    nine: String = null;
    ten: String = null;
}